Update dependencies and migrate controller specs to request specs
- Update critical security vulnerabilities:
  - axios: 1.6.8 → 1.11.0 (fixes SSRF vulnerabilities)
  - factory_bot: 5.2.0 → 6.5.4
  - ESLint: 8.x → 9.33.0
  - Prettier: 2.8.8 → 3.6.2

- Migrate all controller specs to request specs for Rails 8 compatibility
  - Rails 8 lazy route loading breaks controller specs
  - Request specs are the recommended approach
  - All 190 tests passing

- Add Rails.application.reload_routes! to request specs for Rails 8

Authored by: Aaron Lippold<lippold@gmail.com>
