+++ package.json
@@ -5,7 +5,7 @@
     "@rails/actioncable": "^8.0.201",
     "@rails/activestorage": "^8.0.201",
     "@rails/ujs": "^7.1.3-4",
-    "axios": "1.6.8",
+    "axios": "1.11.0",
     "bootstrap": "^4.4.1",
     "bootstrap-icons": "^1.13.1",
     "bootstrap-vue": "^2.13.0",
@@ -30,12 +30,12 @@
     "esbuild": "^0.25.0",
     "esbuild-sass-plugin": "^3.3.1",
     "esbuild-vue": "^1.2.2",
-    "eslint": "^8.25.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "9.33.0",
+    "eslint-config-prettier": "10.1.8",
     "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-prettier-vue": "^3.1.0",
     "eslint-plugin-vue": "^9.6.0",
-    "prettier": "^2.3.2",
+    "prettier": "3.6.2",
     "sass": "^1.85.1"
   },
   "scripts": {
